backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sam Liang"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic-settings = "^2.2.1"
opencv-python = "^4.9.0.80"
torch = "^2.3.0"
torchvision = "^0.18.0"
torchaudio = "^2.3.0"
ftfy = "^6.2.0"
regex = "^2024.4.28"
tqdm = "^4.66.4"
pinecone-client = "^4.0.0"
uvicorn = "^0.29.0"
setuptools = "^69.5.1"
clip = {git = "https://github.com/openai/CLIP.git"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
